First implementation of review feature.
Users can hold a segmented text and access the associated context menu to store the raw text and the corresponding info. In the review tab of the main content view, in a future update users will be able to export all their stored review texts to PC (so that it can be imported into anki).